Consolidate AI commit and amend scripts into a single file
SUMMARY:
This commit consolidates the functionality of `aiCommit.bff.ts` and `aiAmend.bff.ts` into a single file, `ai.bff.ts`. The new file provides the same capabilities for generating AI-assisted commit messages and amending commits, but with improved organization and reduced redundancy. The functions `aiCommit` and `aiAmend` are now part of one unified script. The separate files `aiCommit.bff.ts` and `aiAmend.bff.ts` have been deleted as their functionality is now included in `ai.bff.ts`. This change streamlines the codebase and simplifies maintenance.

TEST PLAN:
1. Set the `OPENAI_API_KEY` environment variable for OpenAI API access.
2. Run `ai:commit` to ensure it generates a commit message using the new `ai.bff.ts`.
3. Run `ai:amend` to amend a previous commit and verify the AI-generated message.
4. Test user prompts for committing and submitting pull requests.
5. Confirm no functionality loss from the previous individual scripts.
