+++ infra/bff/friends/ai.bff.ts
@@ -1,4 +1,4 @@
-// infra/bff/friends/aiAmend.bff.ts
+// infra/bff/friends/ai.bff.ts
 import { register } from "infra/bff/bff.ts";
 import {
   runShellCommand,
@@ -9,16 +9,12 @@ import { connectToOpenAi } from "packages/bolt-foundry/bolt-foundry.ts";
 
 const logger = getLogger(import.meta);
 
-export async function aiAmend(_args: string[]): Promise<number> {
-  logger.info(
-    "Running aiAmend to amend previous commit with AI-generated message...",
-  );
-
+async function runPreCommitChecks(): Promise<boolean> {
   // First, run format to ensure code is properly formatted
   logger.info("Running code formatter first...");
   const formatResult = await runShellCommand(["bff", "f"]);
   if (formatResult !== 0) {
-    logger.warn("Formatting encountered issues, but continuing with amend...");
+    logger.warn("Formatting encountered issues, but continuing with commit...");
   }
 
   // Run linting
@@ -26,9 +22,9 @@ export async function aiAmend(_args: string[]): Promise<number> {
   const lintResult = await runShellCommand(["bff", "lint"]);
   if (lintResult !== 0) {
     logger.error(
-      "Linting failed. Please fix the linting issues before amending commit.",
+      "Linting failed. Please fix the linting issues before committing.",
     );
-    return lintResult;
+    return false;
   }
 
   // Run type checking
@@ -36,20 +32,15 @@ export async function aiAmend(_args: string[]): Promise<number> {
   const checkResult = await runShellCommand(["bff", "check"]);
   if (checkResult !== 0) {
     logger.error(
-      "Type checking failed. Please fix the type errors before amending commit.",
+      "Type checking failed. Please fix the type errors before committing.",
     );
-    return checkResult;
+    return false;
   }
 
-  // Check for OpenAI API key
-  const openaiApiKey = Deno.env.get("OPENAI_API_KEY");
+  return true;
+}
 
-  if (!openaiApiKey) {
-    logger.error("OPENAI_API_KEY environment variable is not set");
-    return 1;
-  }
-
-  // Try to get and configure GitHub user info
+async function configureGitHubUser(): Promise<void> {
   try {
     logger.info("Checking GitHub user information...");
     const { stdout: userInfoJson, code: userInfoCode } =
@@ -86,9 +77,207 @@ export async function aiAmend(_args: string[]): Promise<number> {
   } catch (error) {
     logger.warn("Could not retrieve GitHub user info:", error);
   }
+}
+
+async function trackChanges(): Promise<boolean> {
+  // Track all changes with sl add and sl remove
+  logger.info("Tracking changes with sl add and sl remove");
+
+  // Run sl add to track all new and modified files
+  const addResult = await runShellCommand(["sl", "add", "."]);
+  if (addResult !== 0) {
+    logger.error("Failed to add files to sapling tracking");
+    return false;
+  }
+
+  // Run sl remove to track all deleted files
+  const removeResult = await runShellCommand(["sl", "remove", "--after"]);
+  if (removeResult !== 0) {
+    logger.error("Failed to remove deleted files from sapling tracking");
+    return false;
+  }
+
+  return true;
+}
+
+export async function aiCommit(_args: string[]): Promise<number> {
+  logger.info("Running ai:commit to generate commit message with OpenAI...");
+
+  // Run initial checks and setup
+  if (!await runPreCommitChecks()) return 1;
+
+  // Check for OpenAI API key
+  const openaiApiKey = Deno.env.get("OPENAI_API_KEY");
+  if (!openaiApiKey) {
+    logger.error("OPENAI_API_KEY environment variable is not set");
+    return 1;
+  }
+
+  // Configure GitHub user
+  await configureGitHubUser();
+
+  // Track all changes
+  if (!await trackChanges()) return 1;
+
+  // Generate diff file
+  logger.info("Generating diff file");
+  const { stdout: diffOutput, code: diffCode } =
+    await runShellCommandWithOutput(
+      ["sl", "diff"],
+      {},
+      true,
+      true,
+    );
+
+  if (diffCode !== 0) {
+    logger.error("Failed to generate diff");
+    return diffCode;
+  }
+
+  if (!diffOutput.trim()) {
+    logger.warn("No changes detected in diff");
+    return 0;
+  }
+
+  // Connect to OpenAI and create a custom fetch
+  const openAiFetch = connectToOpenAi(openaiApiKey);
+
+  // Prepare the prompt to send to OpenAI
+  const prompt = `
+I need you to analyze the following git diff and create:
+1. A concise but descriptive commit title (one line, max 72 chars)
+2. A detailed commit message that explains what changed and why
+3. A brief test plan section
+
+Format your response EXACTLY like this, with no extra text:
+TITLE: <commit title>
+
+SUMMARY:
+<summary of changes>
+
+TEST PLAN:
+<test plan>
+
+Here is the diff:
+${diffOutput}
+`;
+
+  // Send to OpenAI and get response
+  logger.info("Sending diff to OpenAI...");
+  try {
+    const response = await openAiFetch(
+      "https://api.openai.com/v1/chat/completions",
+      {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({
+          model: "gpt-4o",
+          messages: [
+            {
+              role: "user",
+              content: prompt,
+            },
+          ],
+          temperature: 0.7,
+          max_tokens: 1024,
+        }),
+      },
+    );
+
+    const result = await response.json();
+    const aiResponse = result.choices[0].message.content.trim();
+
+    // Parse response into title and message
+    const titleMatch = aiResponse.match(/TITLE: (.*)/);
+    const title = titleMatch ? titleMatch[1].trim() : "Automated commit";
+
+    // Extract everything after TITLE line for the commit message
+    const message = aiResponse.replace(/TITLE: .*\n/, "").trim();
+
+    // Display the generated commit message
+    logger.info("OpenAI generated the following commit message:");
+    logger.info(`\n${title}\n\n${message}`);
+
+    // Ask user if they want to commit with this message
+    logger.info("\nDo you want to commit with this message? (y/n)");
+
+    const decoder = new TextDecoder();
+    const buffer = new Uint8Array(1);
+    await Deno.stdin.read(buffer);
+    const answer = decoder.decode(buffer).toLowerCase();
+
+    if (answer === "y") {
+      // Create the commit
+      logger.info("Creating commit...");
+      const commitResult = await runShellCommand([
+        "sl",
+        "commit",
+        "-m",
+        `${title}\n\n${message}`,
+      ]);
+
+      if (commitResult !== 0) {
+        logger.error("Failed to create commit");
+        return commitResult;
+      }
+
+      logger.info("Commit created successfully!");
+
+      // Ask if user wants to submit a pull request
+      logger.info("\nDo you want to submit a pull request? (y/n)");
+      await Deno.stdin.read(buffer);
+      const submitPrAnswer = decoder.decode(buffer).toLowerCase();
+
+      if (submitPrAnswer === "y") {
+        logger.info("Submitting pull request...");
+        const submitResult = await runShellCommand([
+          "sl",
+          "submit",
+        ]);
+
+        if (submitResult !== 0) {
+          logger.error("Failed to submit pull request");
+          return submitResult;
+        }
+
+        logger.info("Pull request submitted successfully!");
+      } else {
+        logger.info("Pull request submission skipped.");
+      }
+
+      return 0;
+    } else {
+      logger.info("Commit cancelled.");
+      return 0;
+    }
+  } catch (error) {
+    logger.error("Error communicating with OpenAI:", error);
+    return 1;
+  }
+}
+
+export async function aiAmend(_args: string[]): Promise<number> {
+  logger.info(
+    "Running ai:amend to amend previous commit with AI-generated message...",
+  );
+
+  // Run initial checks and setup
+  if (!await runPreCommitChecks()) return 1;
+
+  // Check for OpenAI API key
+  const openaiApiKey = Deno.env.get("OPENAI_API_KEY");
+  if (!openaiApiKey) {
+    logger.error("OPENAI_API_KEY environment variable is not set");
+    return 1;
+  }
+
+  // Configure GitHub user
+  await configureGitHubUser();
 
   // 1. Get the current commit message
-  logger.info("1. Getting current commit message");
+  logger.info("Getting current commit message");
   const { stdout: currentCommitMessage, code: commitMsgCode } =
     await runShellCommandWithOutput(
       ["sl", "log", "-r", ".", "--template", "{desc}"],
@@ -105,25 +294,11 @@ export async function aiAmend(_args: string[]): Promise<number> {
   logger.info("Current commit message:");
   logger.info(currentCommitMessage);
 
-  // 2. Track all uncommitted changes with sl add and sl remove
-  logger.info("2. Tracking any uncommitted changes with sl add and sl remove");
-
-  // Run sl add to track all new and modified files
-  const addResult = await runShellCommand(["sl", "add", "."]);
-  if (addResult !== 0) {
-    logger.error("Failed to add files to sapling tracking");
-    return addResult;
-  }
-
-  // Run sl remove to track all deleted files
-  const removeResult = await runShellCommand(["sl", "remove", "--after"]);
-  if (removeResult !== 0) {
-    logger.error("Failed to remove deleted files from sapling tracking");
-    return removeResult;
-  }
+  // 2. Track all uncommitted changes
+  if (!await trackChanges()) return 1;
 
   // 3. Get the diff from the current commit
-  logger.info("3. Generating diff for the current commit");
+  logger.info("Generating diff for the current commit");
   const { stdout: commitDiff, code: commitDiffCode } =
     await runShellCommandWithOutput(
       ["sl", "diff", "-c", "."],
@@ -138,7 +313,7 @@ export async function aiAmend(_args: string[]): Promise<number> {
   }
 
   // 4. Get diff for any uncommitted changes
-  logger.info("4. Generating diff for uncommitted changes");
+  logger.info("Generating diff for uncommitted changes");
   const { stdout: uncommittedDiff, code: uncommittedDiffCode } =
     await runShellCommandWithOutput(
       ["sl", "diff"],
@@ -178,10 +353,10 @@ Take into account the existing commit message when generating the new one.
 Format your response EXACTLY like this, with no extra text:
 TITLE: <commit title>
 
-SUMMARY:
+## SUMMARY
 <summary of changes>
 
-TEST PLAN:
+## TEST PLAN
 <test plan>
 
 Here is the existing commit message:
@@ -192,7 +367,7 @@ ${combinedDiff}
 `;
 
   // 8. Send to OpenAI and get response
-  logger.info("5. Sending combined diff to OpenAI...");
+  logger.info("Sending combined diff to OpenAI...");
   try {
     const response = await openAiFetch(
       "https://api.openai.com/v1/chat/completions",
@@ -226,11 +401,11 @@ ${combinedDiff}
     const message = aiResponse.replace(/TITLE: .*\n/, "").trim();
 
     // 10. Display the generated commit message
-    logger.info("6. OpenAI generated the following commit message:");
+    logger.info("OpenAI generated the following commit message:");
     logger.info(`\n${title}\n\n${message}`);
 
     // 11. Ask user if they want to amend with this message
-    logger.info("\n7. Do you want to amend with this message? (y/n)");
+    logger.info("\nDo you want to amend with this message? (y/n)");
 
     const decoder = new TextDecoder();
     const buffer = new Uint8Array(1);
@@ -239,7 +414,7 @@ ${combinedDiff}
 
     if (answer === "y") {
       // 12. Amend the commit, including any uncommitted changes
-      logger.info("8. Amending commit...");
+      logger.info("Amending commit...");
       // Ensure proper escaping of the commit message
       const amendResult = await runShellCommand([
         "sl",
@@ -257,12 +432,12 @@ ${combinedDiff}
       logger.info("Commit amended successfully!");
 
       // Ask if user wants to submit a pull request
-      logger.info("\n9. Do you want to submit a pull request? (y/n)");
+      logger.info("\nDo you want to submit a pull request? (y/n)");
       await Deno.stdin.read(buffer);
       const submitPrAnswer = decoder.decode(buffer).toLowerCase();
 
       if (submitPrAnswer === "y") {
-        logger.info("10. Submitting pull request...");
+        logger.info("Submitting pull request...");
         const submitResult = await runShellCommand([
           "sl",
           "submit",
@@ -290,7 +465,13 @@ ${combinedDiff}
 }
 
 register(
-  "aiAmend",
+  "ai:commit",
+  "Generate commit message with OpenAI based on current changes",
+  aiCommit,
+);
+
+register(
+  "ai:amend",
   "Amend previous commit with an AI-generated message based on the commit's changes",
   aiAmend,
 );
